requirements.txt
pandas
sentence-transformers
scikit-learn
google-generativeai
gradio
joblib
lime
python-dotenv
